Fix legacy deps and Update blocktype index
Building with webpack was incompatible with current node version due to an update in ssl hash. Using legacy provider fixed this issue.

When trying to hit /blog/posts/:slug, met with an error. After debugging, there was a typescript error in the blogpost, where it couldn't index with BlockTypes due to paragraph being missing. Added a paragraph component, which fixed the typeing issue and sent the data to the dom
 main