Add Carousel component and render client
Made a carousel component which takes in the urls generated from the schema. In order for the event listeners to work, the application sent from the server needed to be rendered to the dom. Added a client side webpack config file to generate that.

This file was accessed through the buildhtml, so scripts needed to be added to the body to link to those public files. Also needed a file to generate so created an App.tsx and used that to hydrate the dom.